test: move DAO-constructing tests into the dao package
