+++ src/test/java/io/appform/dropwizard/sharding/dao/ParentChildTest.java
@@ -1,12 +1,13 @@
-package io.appform.dropwizard.sharding.dao.locktest;
+package io.appform.dropwizard.sharding.dao;
 
 import com.google.common.collect.Lists;
-import io.appform.dropwizard.sharding.DBShardingBundleBase;
 import io.appform.dropwizard.sharding.ShardInfoProvider;
 import io.appform.dropwizard.sharding.config.ShardingBundleOptions;
-import io.appform.dropwizard.sharding.dao.MultiTenantRelationalDao;
-import io.appform.dropwizard.sharding.dao.RelationalDao;
 import io.appform.dropwizard.sharding.dao.interceptors.DaoClassLocalObserver;
+import io.appform.dropwizard.sharding.dao.locktest.Category;
+import io.appform.dropwizard.sharding.dao.locktest.ChildAClass;
+import io.appform.dropwizard.sharding.dao.locktest.ChildBClass;
+import io.appform.dropwizard.sharding.dao.locktest.ParentClass;
 import io.appform.dropwizard.sharding.observers.internal.TerminalTransactionObserver;
 import io.appform.dropwizard.sharding.query.QuerySpec;
 import io.appform.dropwizard.sharding.query.QueryUtils;
