+++ src/test/java/io/appform/dropwizard/sharding/dao/ScrollTest.java
@@ -1,7 +1,8 @@
-package io.appform.dropwizard.sharding;
+package io.appform.dropwizard.sharding.dao;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import io.appform.dropwizard.sharding.ShardInfoProvider;
 import io.appform.dropwizard.sharding.config.ShardingBundleOptions;
 import io.appform.dropwizard.sharding.dao.LookupDao;
 import io.appform.dropwizard.sharding.dao.MultiTenantLookupDao;
